+++ lib/steep.rb
@@ -70,6 +70,7 @@ require "steep/drivers/check"
 require "steep/drivers/validate"
 require "steep/drivers/annotations"
 require "steep/drivers/scaffold"
+require "steep/drivers/print_interface"
 
 if ENV["NO_COLOR"]
   Rainbow.enabled = false