+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+* Add `interface` command to print interface built for given type
 * Add `--strict` option for `check` command
 * Fix `scaffold` command for empty class/modules
 * Type check method definition with empty body